Update calculus-derivatives-with-python.ipynb
In the lesson, there is a graph showing "Another function and its Derivative", but when opened in collab and clicked "execute all" it doesn't shows the graph that is in the lesson but the same graph as the previous part just closer (in other interval of x). So i proposed another function and its derivative declared previous to the graph so it really shows "Another function and its Derivative" and not the same function in the previous example.